+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/shared_helpers/shared_helpers_test.go
@@ -13,7 +13,7 @@ const (
 )
 
 func TestReplaceMagicStringWithValue_SimpleCase(t *testing.T) {
-	instruction := kurtosis_instruction.NewInstructionPosition(5, 3)
+	instruction := kurtosis_instruction.NewInstructionPosition(5, 3, "dummyFile")
 	inputStr := instruction.MagicString(ArtifactUUIDSuffix)
 	environment := startosis_executor.NewExecutionEnvironment()
 	testUuid := "test-uuid"
@@ -27,8 +27,8 @@ func TestReplaceMagicStringWithValue_SimpleCase(t *testing.T) {
 }
 
 func TestReplaceMagicStringWithValue_ValidMultipleReplaces(t *testing.T) {
-	instructionA := kurtosis_instruction.NewInstructionPosition(5, 3)
-	instructionB := kurtosis_instruction.NewInstructionPosition(56, 33)
+	instructionA := kurtosis_instruction.NewInstructionPosition(45, 60, "github.com/kurtosis-tech/eth2-module/src/participant_network/prelaunch_data_generator/el_genesis/el_genesis_data_generator.star")
+	instructionB := kurtosis_instruction.NewInstructionPosition(56, 33, "github.com/kurtosis-tech/eth2-module/src/participant_network/prelaunch_data_generator/cl_genesis/cl_genesis_data_generator.star")
 	magicStringA := instructionA.MagicString(ArtifactUUIDSuffix)
 	magicStringB := instructionB.MagicString(ArtifactUUIDSuffix)
 	inputStr := fmt.Sprintf("%v %v %v", magicStringB, magicStringA, magicStringB)
@@ -47,7 +47,7 @@ func TestReplaceMagicStringWithValue_ValidMultipleReplaces(t *testing.T) {
 }
 
 func TestReplaceMagicStringWithValue_MagicStringNotInEnvironment(t *testing.T) {
-	instruction := kurtosis_instruction.NewInstructionPosition(5, 3)
+	instruction := kurtosis_instruction.NewInstructionPosition(5, 3, "dummyFile")
 	magicString := instruction.MagicString(ArtifactUUIDSuffix)
 	emptyEnvironment := startosis_executor.NewExecutionEnvironment()
 	_, err := ReplaceArtifactUuidMagicStringWithValue(magicString, testServiceId, emptyEnvironment)