+++ docs/changelog.md
@@ -1,5 +1,18 @@
 # TBD
 
+### Features
+- Added exec and HTTP request facts
+- Prints out the instruction line, col & filename in the execution error
+- Prints out the instruction line, col & filename in the validation error
+
+# 0.51.7
+
+### Fixes
+- Fixed instruction position to work with nested functions
+
+### Features
+- Instruction position now contains the filename too
+
 # 0.51.6
 
 ### Features